Test-pass robustness
Amy tested interactive mode and found some ways it needed to be made
more robust, mostly input validation.  It turns out, the way the code
was structured could lead to an infinite loop if the integer inputs
weren’t validated.  Fixed that by ignoring input and re-requesting it
when std::can had the fail bit set.

Also, it was confusing to have the printed output be the cache sorted
by key, so I modified it to return the cache state sorted by most
recently used.