Implement the LruCache with a custom doubly-linked list and an unordered_map
The original problem statement has a solution recommending that the
LruCache be implemented with a custom doubly-linked-list that is mapped
by an unordered map.  This solution doesn’t feel nearly as robust or
maintainable.  In particular, modifying the code could easily introduce
a memory leak.  Making the custom doubly-linked list robust and
exception-safe would be a good next step.  I’d prefer to use an STL
data structure (like list) but unfortunately, that probably requires
deeper access to the list internals (or doing something unnatural like
having the unordered_map reference iterators; maybe that would be
acceptable, but I’d have to look into what might invalidate those
iterators…and it would still be brittle).