Changed the approach signficiantly
The approach I was taking before turns out to be flawed.  Previously, I
sorted Busy first, then tentative, then free (and was removing free
from the vector altogether.)  This doesn’t support Tentative falling
before free properly, and invalidates assumptions I was making about
being able to march forward based on start/end time and assume all
previous ranges are done.

Now, instead, I’m calling appendRanges recursively and have a concept
of “filling”.  If the state is Tentative, you call recursively and fill
with the Hashed pattern if you find any overlapping Busy.  (This means
the recursive depth is only 1 — that is, it’s controlled by the size of
the FreeBusy enum.)

There’s a bug with the two overlapping busy appointments, but I wanted
to get this committed prior to fixing that since the change is
significant.