Updated the readme
I thought it would be important to disclaim that even though I’ve used
this question for interview candidates, the code here is absolutely not
at the level of polish I’d expect in an interview.  No candidate would
have time for that.  Most candidates only get to the class design and
an outline of the algorithm, which is understood and intended.